# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VCA API",
    "version": "0.15.0",
    "description": "백엔드가 생성/관리하는 API 계약. 프론트는 orval(npm run gen:api)로 이 스펙에서 타입/훅을 생성한다.\n\n실시간 데이터(카메라 상태 변화, VIP 감지 이벤트, 집계 카운터)는 이 API가 아닌 MQTT로 전달된다 — 채널 경계와 병합 규칙은 vca-mqtt-broker 레포의 SPEC.md 참고. REST는 스냅샷(페이징 목록·이력·그래프)을, MQTT는 그 이후의 변화분을 담당한다.\n\n공통 규약:\n- 모든 응답은 { success, data, message, code } envelope\n- 페이징은 { content, page, size, totalElements } (page는 0부터)\n- 시각은 ISO-8601 UTC, 날짜 파라미터의 기본값은 사이트 로컬(Asia/Singapore) 기준 오늘\n- ID는 문자열: cameraId/locationId는 ^[a-z0-9-]{1,64}$ (MQTT 토픽 경로와 공유)\n- similarity는 0~1 실수 (표시 포맷은 프론트 책임)"
  },
  "servers": [{ "url": "/api" }],
...
  "paths": {
...
    "/portal/cameras/{cameraId}": {
      "parameters": [{ "name": "cameraId", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": {
        "operationId": "portalGetCamera",
        "summary": "카메라 상세 (UV-53, 카메라 상세 모달)",
        "tags": ["portal"],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PortalCameraResponse" } } } },
          "404": { "description": "ADM-4040", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "put": {
        "operationId": "portalUpdateCamera",
        "summary": "카메라 수정 (UV-53) — password 생략 시 기존 유지",
        "tags": ["portal"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PortalCameraRequest" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PortalCameraResponse" } } } } }
      },
      "delete": {
        "operationId": "portalDeleteCamera",
        "summary": "카메라 삭제 (UV-53) — 모듈 분석 중단 + retained 삭제(v1.9 규칙) + 미디어 path 삭제",
        "tags": ["portal"],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthOkResponse" } } } } }
      }
    },
    "/portal/cameras/bulk/zone": {
      "put": {
        "operationId": "portalSetCamerasZone",
        "summary": "일괄 존 변경 (UV-53) — 감사 로그 1줄(하나의 결정)",
        "tags": ["portal"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PortalBulkZoneRequest" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PortalBulkResultResponse" } } } } }
      }
    },
    "/portal/cameras/bulk/delete": {
      "post": {
        "operationId": "portalDeleteCameras",
        "summary": "일괄 삭제 (UV-53) — 감사 로그 1줄. DELETE 본문 호환성 때문에 POST",
        "tags": ["portal"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PortalBulkDeleteRequest" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PortalBulkResultResponse" } } } } }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "PortalCameraRequest": {
        "type": "object",
        "description": "등록·수정 공용. password 쓰기 전용(수정 시 생략=유지). projectId 생략 시 기본 프로젝트, code 생략 시 CAM-{ZONE3}-{NNN}, zone 생략 시 name, locationId 생략 시 loc-{슬러그}",
        "required": ["name", "rtspUrl", "coordinates"],
        "properties": {
          "projectId": { "type": "string", "nullable": true },
          "code": { "type": "string", "nullable": true, "description": "화면 표시 코드(CAM-BDK-001) — cameraId(내부 키)와 별개" },
          "name": { "type": "string" },
          "ip": { "type": "string", "nullable": true },
          "mac": { "type": "string", "nullable": true },
          "maker": { "type": "string", "nullable": true },
          "model": { "type": "string", "nullable": true },
          "resolution": { "type": "string", "nullable": true, "example": "FHD (1920×1080)" },
          "protocol": { "type": "string", "nullable": true, "enum": ["TCP", "UDP", null] },
          "username": { "type": "string", "nullable": true },
          "password": { "type": "string", "format": "password", "nullable": true },
          "rtspUrl": { "type": "string", "description": "rtsp:// 또는 rtsps:// — 민감정보, 관리자 세션 화면에만" },
          "locationId": { "type": "string", "nullable": true, "pattern": "^[a-z0-9-]{1,64}$" },
          "zone": { "type": "string", "nullable": true, "description": "구역명(Bedok) — 필터·일괄 변경 단위" },
          "location": { "type": "string", "nullable": true, "description": "설치 위치 설명(Control Room, Singapore)" },
          "coordinates": { "$ref": "#/components/schemas/LatLng" },
          "serverId": { "type": "string", "nullable": true, "description": "처리 서버(PortalServer.id) — 없으면 미배정" },
          "aiFeatures": { "type": "array", "items": { "type": "string", "enum": ["Re-ID Analysis", "License Plate Recognition"] } },
          "thumbnail": { "type": "string", "nullable": true }
        }
      },
      "PortalCamera": {
        "type": "object",
        "description": "status/lastSeenAt/lastChangeAt은 원장이 아니라 모듈 MQTT status 적재값 — unknown이면 미수신",
        "required": ["cameraId", "name", "hasCredential", "rtspUrl", "locationId", "zone", "coordinates", "aiFeatures", "status", "createdAt", "updatedAt"],
        "properties": {
          "cameraId": { "type": "string", "pattern": "^[a-z0-9-]{1,64}$" },
          "projectId": { "type": "string", "nullable": true },
          "code": { "type": "string", "nullable": true },
          "name": { "type": "string" },
          "ip": { "type": "string", "nullable": true },
          "mac": { "type": "string", "nullable": true },
          "maker": { "type": "string", "nullable": true },
          "model": { "type": "string", "nullable": true },
          "resolution": { "type": "string", "nullable": true },
          "protocol": { "type": "string", "nullable": true },
          "username": { "type": "string", "nullable": true },
          "hasCredential": { "type": "boolean" },
          "rtspUrl": { "type": "string" },
          "locationId": { "type": "string" },
          "zone": { "type": "string" },
          "location": { "type": "string", "nullable": true },
          "coordinates": { "$ref": "#/components/schemas/LatLng" },
          "serverId": { "type": "string", "nullable": true },
          "aiFeatures": { "type": "array", "items": { "type": "string" } },
          "thumbnail": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["online", "offline", "error", "unknown"] },
          "lastSeenAt": { "type": "string", "format": "date-time", "nullable": true, "description": "마지막 status 수신" },
          "lastChangeAt": { "type": "string", "format": "date-time", "nullable": true, "description": "마지막 상태 전이" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "PortalBulkZoneRequest": {
        "type": "object",
        "required": ["cameraIds", "zone"],
        "properties": { "cameraIds": { "type": "array", "items": { "type": "string" } }, "zone": { "type": "string" } }
      },
      "PortalBulkDeleteRequest": {
        "type": "object",
        "required": ["cameraIds"],
        "properties": { "cameraIds": { "type": "array", "items": { "type": "string" } } }
      },
      "PortalBulkResult": {
        "type": "object",
        "required": ["affected", "cameraIds"],
        "properties": { "affected": { "type": "integer" }, "cameraIds": { "type": "array", "items": { "type": "string" } } }
      },
...
      "PortalCameraResponse": { "type": "object", "required": ["success", "code", "data"], "properties": { "success": { "type": "boolean" }, "code": { "type": "string" }, "message": { "type": "string", "nullable": true }, "data": { "$ref": "#/components/schemas/PortalCamera" } } },
...
      "PortalBulkResultResponse": { "type": "object", "required": ["success", "code", "data"], "properties": { "success": { "type": "boolean" }, "code": { "type": "string" }, "message": { "type": "string", "nullable": true }, "data": { "$ref": "#/components/schemas/PortalBulkResult" } } },
...
      "AuthOkResponse": {
        "type": "object",
        "required": ["success", "code"],
        "properties": {
          "success": { "type": "boolean" },
          "code": { "type": "string" },
          "message": { "type": "string", "nullable": true },
          "data": { "nullable": true }
        }
      },
...
      "LatLng": {
        "type": "object",
        "description": "WGS84 좌표",
        "required": ["lat", "lng"],
        "properties": {
          "lat": { "type": "number", "format": "double" },
          "lng": { "type": "number", "format": "double" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "code"],
        "properties": {
          "success": { "type": "boolean", "example": false },
          "code": { "type": "string", "description": "오류 코드 (예: VCA-4041)" },
          "message": { "type": "string", "nullable": true, "description": "사람이 읽는 오류 사유" },
          "data": { "nullable": true, "description": "오류 시 null" }
        }
      },
...
    }
  }
}